AI_PaperIQ_Streamlit/requirements.txt
streamlit
google-generativeai
python-dotenv
requests
PyMuPDF
nltk
scikit-learn
plotly
textstat
pandas